Add files via upload
image_listener.py file is used to capture the images infront of cat robot, if it detects other yellow robot in front of it, it sends coordinates to y_cordinate.py file. y_cordinate.py file is used to track yellow robot. I have used turtlesim robot to check working of this node.